// packages/ui/src/components/ui/form_disabled/form.types.ts
export type FormMessageType = 'error' | 'warning' | 'success' | 'info'
// ...
export interface FormMessageProps {
  type?: FormMessageType
  class?: string
  icon?: JSX.Element
  children: JSX.Element
}